requirements.txt.txt
numpy==1.26.4
pandas==2.1.3
streamlit==1.29.0
tensorflow==2.14.0
Pillow==10.0.0
PyMuPDF==1.24.1